Fix comments
