+++ src/app/core/interfaces/course.ts
