+++ src/app/core/services/auth.service.ts
@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {IUser} from '../interfaces/user';
-import localStorageExp from '../../core/helpers/localstorage/localStorageExp';
+import { LocalStorageService } from '../../core/services/local-storage.service';
 
 @Injectable({
   providedIn: 'root'
@@ -10,7 +10,7 @@ export class AuthService {
   private USER_KEY: string;
   private fakeUser: IUser;
 
-  constructor() {
+  constructor(private localStorageService: LocalStorageService ) {
     this.USER_KEY = 'USER_KEY';
     this.fakeUser = {
       id: 1,
@@ -18,25 +18,25 @@ export class AuthService {
       last_name: 'Osipov',
       login: 'test_login',
       password: 'test-pass',
-    }
+    };
   }
 
   login(user: IUser): void {
     console.log('login action with creds: ', user.login, user.password);
-    localStorageExp.setItem(this.USER_KEY, this.fakeUser);
+    this.localStorageService.setItem(this.USER_KEY, this.fakeUser);
   }
 
   logout(): void {
     console.log('logout');
-    localStorageExp.removeItem(this.USER_KEY, this.fakeUser);
+    this.localStorageService.removeItem(this.USER_KEY);
   }
 
   isAuthenticated(): boolean {
-    return !!localStorageExp.getItem(this.USER_KEY);
+    return !!this.localStorageService.getItem(this.USER_KEY);
   }
 
   getUserInfo(): IUser {
-    return localStorageExp.getItem(this.USER_KEY);
+    return this.localStorageService.getItem(this.USER_KEY);
 
   }
 }
